WIP - begin ClrObject refactor
