+++ src/Microsoft.Diagnostics.Runtime/src/Builders/RcwHelpers.cs
@@ -5,19 +5,20 @@
 using System.Collections.Immutable;
 using Microsoft.Diagnostics.Runtime.DacInterface;
 using Microsoft.Diagnostics.Runtime.Implementation;
+using Microsoft.Diagnostics.Runtime;
 
 namespace Microsoft.Diagnostics.Runtime.Builders
 {
-    internal class RcwBuilder : IRcwData
+    internal class RcwHelpers : IRcwHelpers
     {
         private RcwData _rcwData;
         private readonly SOSDac _sos;
-        private readonly RuntimeBuilder _builder;
+        private readonly IClrTypeFactory _factory;
 
-        public RcwBuilder(SOSDac sos, RuntimeBuilder builder)
+        public RcwHelpers(SOSDac sos, IClrTypeFactory factory)
         {
             _sos = sos;
-            _builder = builder;
+            _factory = factory;
         }
 
         public bool Init(ulong obj)
@@ -34,20 +35,17 @@ namespace Microsoft.Diagnostics.Runtime.Builders
 
         public ulong Address { get; private set; }
 
-        ulong IRcwData.IUnknown => _rcwData.IUnknownPointer;
-        ulong IRcwData.VTablePointer => _rcwData.VTablePointer;
-        int IRcwData.RefCount => _rcwData.RefCount;
-        ulong IRcwData.ManagedObject => _rcwData.ManagedObject;
-        bool IRcwData.Disconnected => _rcwData.IsDisconnected != 0;
-        ulong IRcwData.CreatorThread => _rcwData.CreatorThread;
+        ulong IRcwHelpers.IUnknown => _rcwData.IUnknownPointer;
+        ulong IRcwHelpers.VTablePointer => _rcwData.VTablePointer;
+        int IRcwHelpers.RefCount => _rcwData.RefCount;
+        ulong IRcwHelpers.ManagedObject => _rcwData.ManagedObject;
+        bool IRcwHelpers.Disconnected => _rcwData.IsDisconnected != 0;
+        ulong IRcwHelpers.CreatorThread => _rcwData.CreatorThread;
 
-        ImmutableArray<ComInterfaceData> IRcwData.GetInterfaces()
+        ImmutableArray<ComInterfaceData> IRcwHelpers.GetInterfaces()
         {
-            COMInterfacePointerData[]? ifs = _sos.GetRCWInterfaces(Address, _rcwData.InterfaceCount);
-            if (ifs is null)
-                return ImmutableArray<ComInterfaceData>.Empty;
-
-            return _builder.CreateComInterfaces(ifs);
+            return _factory.GetRCWInterfaces(Address, _rcwData.InterfaceCount);
         }
+
     }
 }
