+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/ITypeFactory.cs
@@ -11,23 +11,7 @@ namespace Microsoft.Diagnostics.Runtime.Implementation
 {
     internal interface ITypeFactory : IDisposable
     {
-        bool IsThreadSafe { get; }
-
         ClrRuntime GetOrCreateRuntime();
         ClrHeap GetOrCreateHeap();
-        ClrModule GetOrCreateModule(ClrAppDomain domain, ulong address);
-        bool CreateMethodsForType(ClrType type, out ImmutableArray<ClrMethod> methods);
-        bool CreateFieldsForType(ClrType type, out ImmutableArray<ClrInstanceField> fields, out ImmutableArray<ClrStaticField> staticFields);
-        ComCallableWrapper? CreateCCWForObject(ulong obj);
-        RuntimeCallableWrapper? CreateRCWForObject(ulong obj);
-        ClrType CreateSystemType(ClrHeap heap, ulong mt, string kind);
-        ClrType? GetOrCreateType(ClrHeap heap, ulong mt, ulong obj);
-        ClrType? GetOrCreateType(ulong mt, ulong obj);
-        ClrType GetOrCreateBasicType(ClrElementType basicType);
-        ClrType? GetOrCreateArrayType(ClrType inner, int ranks);
-        ClrType? GetOrCreateTypeFromToken(ClrModule module, int token);
-        ClrType? GetOrCreateTypeFromSignature(ClrModule? module, SigParser parser, IEnumerable<ClrGenericParameter> typeParameters, IEnumerable<ClrGenericParameter> methodParameters);
-        ClrType? GetOrCreatePointerType(ClrType innerType, int depth);
-        ClrMethod? CreateMethodFromHandle(ulong methodHandle);
     }
 }
